Add Easy Java Solution from LeetCode

+++ LeetCode/Java/Easy/1950. Sign of the Product of an Array/Solution.java
@@ -1,6 +1,6 @@
 class Solution {
     public int arraySign(int[] nums) {
-        int pro = 1;
+        double pro = 1;
         for(int i = 0 ; i < nums.length ; i++){
             pro = pro * nums[i];
         }
